Remove AQI and PM2.5 stuff.
Remove chart for PM2.5 data. The space used for the AQI calculation based on the PM2.5 reading will be replaced with the 24 hour PSI instead.

This commit removes the code updating the AQI number display and renames the space for PSI 24 display. 24 hour PSI data isn't in yet.

Fixes fonglh/psi-plot#1.