Refactor. Move duplicate PSI 3 hour update code into a separate function.
Common elements for initializing and updating the PSI 3 hour display were duplicated in the chart creation and update functions.

Move the code for updating the PSI value, descriptor, background color and date to a common function which is then called on initial page load and when dynamically updating the value. DRY